Docker_ETL/etl_job/requirements.txt
pymongo==3.12.0
SQLAlchemy==1.4.15
psycopg2-binary==2.8.6
pandas==1.3.5
regex==2022.1.18
vaderSentiment==3.3.2

